Adding a new storage backend.
At first the fork was a make a XvBA driver for VDPAU.
Taking out one of the two middle mans in the pipeline.

But after I realize that VDPAU has a lot more functionality.
In order to grasp all the aspect in the code I started to develop a
simple handle storage. This will in the long run clear one extra
dependency (GLib).

There is also a performance improvement.
Instead of using for loops for memory copy we can use memcpy().

The next step is to divide de code of vdpau-soft into several modules as
represented in the documentation. This way it will be easier to
understand the code and improve/modify certain modules.